Put all theme service color manipulation methods into a separate typescript class. Updated all components that were calling those methods to use the new theme-utils class
